code split up into more modular bits
